UPSTREAM: fuse: realloc page array
Writeback caching currently allocates requests with the maximum number of
possible pages, while the actual number of pages per request depends on a
couple of factors that cannot be determined when the request is allocated
(whether page is already under writeback, whether page is contiguous with
previous pages already added to a request).

This patch allows such requests to start with no page allocation (all pages
inline) and grow the page array on demand.

If the max_pages tunable remains the default value, then this will mean
just one allocation that is the same size as before.  If the tunable is
larger, then this adds at most 3 additional memory allocations (which is
generously compensated by the improved performance from the larger
request).

Change-Id: I75f64c125fcd899848f0acd86731c057899e3c5e
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: rajdeep-3305 <rajdeepbiswas3305@gmail.com>