Makefile: Switch to -O3 optimization level
-O3 is much more stable with modern compilers these days than it was a
decade ago. Using -O3 on the kernel results in significantly improved
hackbench performance, which is a sign that overall performance in the
kernel is improved. It works especially well in conjunction with LTO.

Tune it for our cpu. This yields a modest, consistent improvement in hackbench performance. It
provides the greatest performance improvement among all the relevant	-mcpu
options available.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adithya R <radithya2002@gmail.com>
Signed-off-by: danascape <saalimquadri1@gmail.com>
Signed-off-by: MadhavSaladi <baaswanthmadhav@gmail.com>
Signed-off-by: techyminati <sinha.aryan03@gmail.com>
Signed-off-by: Aeron-Aeron <aeronrules2@gmail.com>
Signed-off-by: rajdeep-3305 <rajdeepbiswas3305@gmail.com>